requirements.txt
aiomysql==0.2.0
anyio==4.2.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.94.0
fastapi-extend
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
pydantic==1.9.0
PyJWT==2.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
simplejson==3.19.2
sniffio==1.3.0
SQLAlchemy==2.0.5
starlette==0.26.1
typing_extensions==4.9.0
uvicorn==0.25.0
